+++ packages/experiments/src/components/Shimmer/Shimmer.base.tsx
@@ -32,11 +32,22 @@ export class ShimmerBase extends BaseComponent<IShimmerProps, {}> {
   }
 
   public render(): JSX.Element {
-    const { getStyles, width, lineElements, children, isDataLoaded, isBaseStyle } = this.props;
+    const {
+      getStyles,
+      width,
+      lineElements,
+      children,
+      isDataLoaded,
+      isBaseStyle,
+      widthInPercentage,
+      widthInPixel
+    } = this.props;
 
     const rowHeight: number | undefined = lineElements ? findMaxElementHeight(lineElements) : undefined;
 
-    this._classNames = getClassNames(getStyles!, { width, rowHeight, isDataLoaded, isBaseStyle });
+    this._classNames = getClassNames(getStyles!, {
+      width, rowHeight, isDataLoaded, isBaseStyle, widthInPercentage, widthInPixel
+    });
 
     const renderedElements: React.ReactNode = getRenderedElements(lineElements, rowHeight);