+++ packages/styling/src/styles/getFocusStyle.ts
@@ -2,6 +2,7 @@ import { IRawStyle } from '@uifabric/merge-styles/lib/index';
 import { ITheme } from '../interfaces/index';
 import { HighContrastSelector } from './CommonStyles';
 import { ZIndexes } from './zIndexes';
+import { IsFocusVisibleClassName } from '@uifabric/utilities/lib/index';
 
 /**
  * Generates a focus style which can be used to define an :after focus border.
@@ -30,7 +31,7 @@ export function getFocusStyle(
         border: '0'
       },
 
-      '.ms-Fabric.is-focusVisible &:focus:after': {
+      [`.${IsFocusVisibleClassName} &:focus:after`]: {
         content: '""',
         position: 'absolute',
         left: inset + 1,